Move more property auth tests
